Move mars into its own file
