--- openfisca_us/variables/demographic/tax_unit/MARSType.py
@@ -1,9 +1,0 @@
-from openfisca_us.model_api import *
-
-
-class MARSType(Enum):
-    SINGLE = "Single"
-    JOINT = "Joint"
-    SEPARATE = "Separate"
-    HEAD_OF_HOUSEHOLD = "Head of household"
-    WIDOW = "Widow(er)"
